% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{check_package_functions}
\alias{check_package_functions}
\title{Check a Package for Missing Functions}
\usage{
check_package_functions(api = c("ec2", "lambda"))
}
\arguments{
\item{api}{The api to check.}
}
\value{
A list with components \code{missing_functions} (functions that
  should be present based on the actions in the API reference),
  \code{unmatched_functions} (functions that are exported by the package but
  not matched in the API reference), and \code{matched_functions} (functions
  that are just right).
}
\description{
Check an aws.* package for missing functions and unmatched functions.
}
\examples{
check_package_functions("lambda")
}
