% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actions.R
\name{get_actions}
\alias{get_actions}
\title{Get API Actions}
\usage{
get_actions(api = c("ec2", "lambda"))
}
\arguments{
\item{api}{The name of the API for which data should be gathered.}
}
\value{
A character vector of actions
}
\description{
Get the names of all actions for a given AWS API.
}
\examples{
get_actions("ec2")
}
