% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoints.R
\name{get_endpoints}
\alias{get_endpoints}
\title{Get API Endpoints}
\usage{
get_endpoints(api = c("ec2", "lambda"))
}
\arguments{
\item{api}{The name of the API for which data should be gathered.}
}
\value{
A character vector of endpoints.
}
\description{
Get the names of all endpoints for a given AWS API.
}
\examples{
get_endpoints("ec2")
}
