% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_expected_functions}
\alias{get_expected_functions}
\title{Get Expected Function Names}
\usage{
get_expected_functions(actions)
}
\arguments{
\item{actions}{A character vector of action names in CamelCase, as returned
by \code{\link{get_actions}}.}
}
\value{
A character vector of the same length as \code{actions}.
}
\description{
Convert action names (CamelCase) to function names (snake_case).
}
\examples{
get_expected_functions(
  c("This", "ThatOne", "AnotherActionName")
)
}
